requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
supabase==2.4.0
python-multipart==0.0.6
gunicorn==21.2.0
